producer/requirements.txt
kafka-python-ng==2.2.3
pandas==2.2.3
